+++ openapiexplorer-20241130/ChangeLog.md
@@ -1,3 +1,27 @@
+2025-12-09 Version: 1.2.0
+- Support API GetApiMcpServerUserConfig.
+- Support API UpdateApiMcpServerUserConfig.
+- Update API CreateApiMcpServer: add request parameters body.enableCustomVpcWhitelist.
+- Update API CreateApiMcpServer: add request parameters body.publicAccess.
+- Update API CreateApiMcpServer: add request parameters body.vpcWhitelists.
+- Update API CreateApiMcpServer: add response parameters Body.urls.vpcMcp.
+- Update API CreateApiMcpServer: add response parameters Body.urls.vpcSse.
+- Update API GenerateCLICommand: add request parameters jsonApiParams.
+- Update API GetApiMcpServer: add response parameters Body.enableCustomVpcWhitelist.
+- Update API GetApiMcpServer: add response parameters Body.publicAccess.
+- Update API GetApiMcpServer: add response parameters Body.vpcWhitelists.
+- Update API GetApiMcpServer: add response parameters Body.urls.vpcMcp.
+- Update API GetApiMcpServer: add response parameters Body.urls.vpcSse.
+- Update API ListApiMcpServers: add response parameters Body.apiMcpServers.$.enableCustomVpcWhitelist.
+- Update API ListApiMcpServers: add response parameters Body.apiMcpServers.$.publicAccess.
+- Update API ListApiMcpServers: add response parameters Body.apiMcpServers.$.vpcWhitelists.
+- Update API ListApiMcpServers: add response parameters Body.apiMcpServers.$.urls.vpcMcp.
+- Update API ListApiMcpServers: add response parameters Body.apiMcpServers.$.urls.vpcSse.
+- Update API UpdateApiMcpServer: add request parameters body.enableCustomVpcWhitelist.
+- Update API UpdateApiMcpServer: add request parameters body.publicAccess.
+- Update API UpdateApiMcpServer: add request parameters body.vpcWhitelists.
+
+
 2025-11-10 Version: 1.1.0
 - Support API ApiMcpServerValidateHcl.
 - Support API CreateApiMcpServer.